Move event-notification-types out of form
